31 database design (#33)
* Database design, created database entities, created associations

* Database; created associations including dish types

* Database design completed (Closes: #31)

+++ app/src/main/java/com/diegomfv/splendidrecipesmvvm/data/model/dto/RandomRecipesResult.kt
@@ -12,7 +12,7 @@ data class RandomRecipesResult(
 
 @Parcelize
 data class RecipeServerDTO(
-    @SerializedName("id") val id: Int?,
+    @SerializedName("id") val id: Long?,
     @SerializedName("title") val title: String?,
     @SerializedName("image") val image: String?,
     @SerializedName("imageType") val imageType: String?,
